Beginning to separate out pointer scan compontents
