+++ Squalr/Source/Scanners/Pointers/PointerRetracer.cs
@@ -1,0 +1,225 @@
+﻿namespace Squalr.Source.Scanners.Pointers
+{
+    using Snapshots;
+    using SqualrCore.Source.ActionScheduler;
+    using System;
+    using System.Collections.Generic;
+    using System.Threading;
+
+    /// <summary>
+    /// Class to collect all pointers in the running process.
+    /// </summary>
+    internal class PointerRetracer : ScheduledTask
+    {
+        /// <summary>
+        /// Time in milliseconds between scans.
+        /// </summary>
+        private const Int32 RescanTime = 256;
+
+        /// <summary>
+        /// The rounding size for pointer destinations.
+        /// </summary>
+        private const Int32 ChunkSize = 1024;
+
+        /// <summary>
+        /// Gets or sets the number of regions processed by this prefilter.
+        /// </summary>
+        private Int64 processedCount;
+
+        /// <summary>
+        /// Creates an instance of the <see cref="PointerRetracer" /> class.
+        /// </summary>
+        public PointerRetracer() : base(
+            taskName: "Pointer Retracer",
+            isRepeated: false,
+            trackProgress: false)
+        {
+            this.AccessLock = new Object();
+
+            this.Dependencies.Enqueue(new PointerBackTracer());
+        }
+
+        /// <summary>
+        /// Gets or sets the current snapshot being parsed.
+        /// </summary>
+        private Snapshot CurrentSnapshot { get; set; }
+
+        private Object AccessLock { get; set; }
+
+        /// <summary>
+        /// Gets all found pointers in the external process.
+        /// </summary>
+        /// <returns>A set of all found pointers.</returns>
+        public IEnumerable<IntPtr> GetFoundPointerDestinations()
+        {
+            return null;
+        }
+
+        protected override void OnBegin()
+        {
+            this.CurrentSnapshot = SnapshotManager.GetInstance().CreateSnapshotFromUsermodeMemory();
+        }
+
+        /// <summary>
+        /// Called when the scan updates.
+        /// </summary>
+        /// <param name="cancellationToken">The cancellation token for handling canceled tasks.</param>
+        protected override void OnUpdate(CancellationToken cancellationToken)
+        {
+            /*
+            if (this.ScanConstraintManager == null || this.ScanConstraintManager.Count() <= 0)
+            {
+                return;
+            }
+
+            if (this.AcceptedPointers == null || this.AcceptedPointers.Count == 0)
+            {
+                return;
+            }
+
+            List<IntPtr> resolvedAddresses = new List<IntPtr>();
+            List<SnapshotRegionDeprecating> regions = new List<SnapshotRegionDeprecating>();
+
+            // Resolve addresses
+            foreach (Tuple<IntPtr, List<Int32>> fullPointer in this.AcceptedPointers)
+            {
+                resolvedAddresses.Add(this.ResolvePointer(fullPointer));
+            }
+
+            // Build regions from resolved address
+            foreach (IntPtr pointer in resolvedAddresses)
+            {
+                regions.Add(new SnapshotRegionDeprecating<Null>(pointer, Marshal.SizeOf(this.ScanConstraintManager.ElementType)));
+            }
+
+            // Create a snapshot from regions
+            SnapshotDeprecating<Null> pointerSnapshot = new SnapshotDeprecating<Null>(regions);
+
+            // Read the memory (collecting values)
+            pointerSnapshot.ReadAllSnapshotMemory();
+            pointerSnapshot.ElementType = this.ScanConstraintManager.ElementType;
+            this.Snapshot.Alignment = sizeof(Int32);
+            pointerSnapshot.MarkAllValid();
+
+            if (pointerSnapshot.GetRegionCount() <= 0)
+            {
+                this.AcceptedPointers = new List<Tuple<IntPtr, List<Int32>>>();
+                return;
+            }
+
+            // Note there are likely only a few regions that span <= 8 bytes, we do not need to parallelize this
+            foreach (SnapshotRegionDeprecating region in pointerSnapshot)
+            {
+                if (!region.HasValues())
+                {
+                    region.MarkAllInvalid();
+                    continue;
+                }
+
+                foreach (SnapshotElementDeprecating element in region)
+                {
+                    // Enforce each value constraint on the element
+                    foreach (ScanConstraint scanConstraint in this.ScanConstraintManager)
+                    {
+                        switch (scanConstraint.Constraint)
+                        {
+                            case ConstraintsEnum.Equal:
+                                if (!element.EqualToValue(scanConstraint.ConstraintValue))
+                                {
+                                    element.Valid = false;
+                                }
+
+                                break;
+                            case ConstraintsEnum.NotEqual:
+                                if (!element.NotEqualToValue(scanConstraint.ConstraintValue))
+                                {
+                                    element.Valid = false;
+                                }
+
+                                break;
+                            case ConstraintsEnum.GreaterThan:
+                                if (!element.GreaterThanValue(scanConstraint.ConstraintValue))
+                                {
+                                    element.Valid = false;
+                                }
+
+                                break;
+                            case ConstraintsEnum.GreaterThanOrEqual:
+                                if (!element.GreaterThanOrEqualToValue(scanConstraint.ConstraintValue))
+                                {
+                                    element.Valid = false;
+                                }
+
+                                break;
+                            case ConstraintsEnum.LessThan:
+                                if (!element.LessThanValue(scanConstraint.ConstraintValue))
+                                {
+                                    element.Valid = false;
+                                }
+
+                                break;
+                            case ConstraintsEnum.LessThanOrEqual:
+                                if (!element.LessThanOrEqualToValue(scanConstraint.ConstraintValue))
+                                {
+                                    element.Valid = false;
+                                }
+
+                                break;
+                            case ConstraintsEnum.NotScientificNotation:
+                                if (element.IsScientificNotation())
+                                {
+                                    element.Valid = false;
+                                }
+
+                                break;
+                            default:
+                                throw new Exception("Invalid Constraint");
+                        }
+                    }
+                    //// End foreach Constraint
+                }
+                //// End foreach Element
+            }
+            //// End foreach Region
+
+            pointerSnapshot.DiscardInvalidRegions();
+
+            List<Tuple<IntPtr, List<Int32>>> retainedPointers = new List<Tuple<IntPtr, List<Int32>>>();
+
+            if (pointerSnapshot.GetRegionCount() <= 0)
+            {
+                this.AcceptedPointers = retainedPointers;
+                return;
+            }
+
+            // Keep all remaining pointers
+            foreach (SnapshotRegionDeprecating region in pointerSnapshot)
+            {
+                foreach (SnapshotElementDeprecating element in region)
+                {
+                    for (Int32 addressIndex = 0; addressIndex < resolvedAddresses.Count; addressIndex++)
+                    {
+                        if (resolvedAddresses[addressIndex] != element.BaseAddress)
+                        {
+                            continue;
+                        }
+
+                        retainedPointers.Add(this.AcceptedPointers[addressIndex]);
+                    }
+                }
+            }
+
+            this.AcceptedPointers = retainedPointers;
+            */
+        }
+
+        /// <summary>
+        /// Called when the repeated task completes.
+        /// </summary>
+        protected override void OnEnd()
+        {
+        }
+    }
+    //// End class
+}
+//// End namespace
